refactor: rename mobilecalendar to mobile-calendar
